replace `updateLocalSchemaRegistryURI()` with more usage of `discoverSchemaRegistry()` when updating local connections based on sidecar knowledge
